Use true buffer size
A SQL Server odbc driver regression occured because it needed a buffer
with space for 2 null terminator bytes and this code cause the buffer to
need 3 null terminator bytes. ODBC assume the buffer size contains space
for the null terminator already and don't write above the provided buffer
size, so SIZE-1 is not required.